+++ Sources/Models/MessageKitError.swift
@@ -1,18 +1,18 @@
 /*
  MIT License
- 
- Copyright (c) 2017-2018 MessageKit
- 
+
+ Copyright (c) 2017 MessageKit
+
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
  in the Software without restriction, including without limitation the rights
  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  copies of the Software, and to permit persons to whom the Software is
  furnished to do so, subject to the following conditions:
- 
+
  The above copyright notice and this permission notice shall be included in all
  copies or substantial portions of the Software.
- 
+
  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -22,18 +22,15 @@
  SOFTWARE.
  */
 
-import Foundation
-
-extension UIColor {
-
-    static let incomingGray = UIColor(red: 230/255, green: 230/255, blue: 235/255, alpha: 1.0)
-
-    static let outgoingGreen = UIColor(red: 69/255, green: 214/255, blue: 93/255, alpha: 1.0)
-
-    static let inputBarGray = UIColor(red: 247/255, green: 247/255, blue: 247/255, alpha: 1.0)
-
-    static let playButtonLightGray = UIColor(red: 230/255, green: 230/255, blue: 230/255, alpha: 1.0)
-
-    static let sendButtonBlue = UIColor(red: 15/255, green: 135/255, blue: 255/255, alpha: 1.0)
-
+enum MessageKitError {
+    static let avatarPositionUnresolved = "AvatarPosition Horizontal.natural needs to be resolved."
+    static let nilMessagesDataSource = "MessagesDataSource has not been set."
+    static let nilMessagesDisplayDelegate = "MessagesDisplayDelegate has not been set."
+    static let nilMessagesLayoutDeleagte = "MessagesLayoutDelegate has not been set."
+    static let notMessagesCollectionView = "The collectionView is not a MessagesCollectionView."
+    static let layoutUsedOnForeignType = "MessagesCollectionViewFlowLayout is being used on a foreign type."
+    static let unrecognizedSectionKind = "Received unrecognized element kind:"
+    static let unrecognizedCheckingResult = "Received an unrecognized NSTextCheckingResult.CheckingType"
+    static let couldNotLoadAssetsBundle = "MessageKit: Could not load the assets bundle"
+    static let couldNotCreateAssetsPath = "MessageKit: Could not create path to the assets bundle."
 }