+++ Sources/Views/MessageLabel.swift
@@ -1,7 +1,7 @@
 /*
  MIT License
 
- Copyright (c) 2017 MessageKit
+ Copyright (c) 2017-2018 MessageKit
 
  Permission is hereby granted, free of charge, to any person obtaining a copy
  of this software and associated documentation files (the "Software"), to deal
@@ -245,7 +245,7 @@ open class MessageLabel: UILabel {
         let mutableAttributedString = NSMutableAttributedString(attributedString: attributedText)
 
         for detector in detectors {
-            guard let rangeTuples = rangesForDetectors[detector] else { return }
+            guard let rangeTuples = rangesForDetectors[detector] else { continue }
 
             for (range, _)  in rangeTuples {
                 let attributes = detectorAttributes(for: detector)
@@ -283,7 +283,7 @@ open class MessageLabel: UILabel {
         case .link:
             return urlAttributes
         default:
-            fatalError("Received an unrecognized NSTextCheckingResult.CheckingType")
+            fatalError(MessageKitError.unrecognizedCheckingResult)
         }
     }
 
@@ -412,7 +412,7 @@ open class MessageLabel: UILabel {
     
 }
 
-fileprivate enum MessageTextCheckingType {
+private enum MessageTextCheckingType {
     case addressComponents([NSTextCheckingKey: String]?)
     case date(Date?)
     case phoneNumber(String?)